Separate entrypoint from definition of DemolitionSpider
